requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
django-rest-knox>=4.2.0
Pillow>=10.0.0
cryptography>=41.0.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7